+++ corpus/macro-from-source/src/usage.clj
@@ -1,0 +1,5 @@
+(ns usage
+  (:require [script]))
+
+(script/my-let [x 1] (inc x))
+(script/shout "hello")